Incorporated fmt library
commit d65b9422952e09f2cf67598b38ba991028242681
Author: lriki <lriki.net@gmail.com>
Date:   Fri Aug 12 09:53:00 2022 +0900

    wip

commit b855e9ba06a7b5da9eca4612f23e4cd3f94f536c
Author: lriki <lriki.net@gmail.com>
Date:   Fri Aug 12 08:58:05 2022 +0900

    wip

commit 3372aa1616560bd712e966ed31f4e852b332d04a
Author: lriki <lriki.net@gmail.com>
Date:   Wed Aug 10 20:37:28 2022 +0900

    wip

commit 9b7d846d39caf78acab7e7b65a79fc4c6b2a7f32
Author: lriki <lriki.net@gmail.com>
Date:   Wed Aug 10 12:41:30 2022 +0900

    wip

commit cebca56b09fb2396ecc1c3c698e4361195f1d482
Author: lriki <lriki.net@gmail.com>
Date:   Tue Aug 9 20:21:23 2022 +0900

    wip

commit 01da3294662865c1df642b75a1f876ee7ed15567
Author: lriki <lriki.net@gmail.com>
Date:   Mon Aug 8 19:48:50 2022 +0900

    wip

commit e4aa8e8a0c32eefdef85422b5a00fd0c6caa8e0f
Author: lriki <lriki.net@gmail.com>
Date:   Thu Aug 4 17:46:54 2022 +0900

    wip

commit b54b6ffeddfd483ad27cfdb6b6873b35f8b41570
Author: lriki <lriki.net@gmail.com>
Date:   Wed Aug 3 18:07:40 2022 +0900

    wip

commit 45850de50c882de906ca43f994d75d3a9dc4d2c3
Author: lriki <lriki.net@gmail.com>
Date:   Wed Aug 3 14:07:37 2022 +0900

    wip

commit cfd20881fd3b17af6d71eeb9a305ceefd1525c8a
Author: lriki <lriki.net@gmail.com>
Date:   Wed Aug 3 12:04:58 2022 +0900

    wip

commit 9bd04e15c6ad9b1dea4a58bf8131003c9a0105fd
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 29 17:51:00 2022 +0900

    wip

commit c6f3114d8d1561bac252bf0d9e1e5f25187517ad
Author: lriki <lriki.net@gmail.com>
Date:   Thu Jul 28 18:18:59 2022 +0900

    wip

commit a5e520b66416c594535544332eea7d8da77082bb
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 27 18:00:28 2022 +0900

    wip

+++ lumino/LuminoCore/src/IO/DllLoader.cpp
@@ -15,10 +15,10 @@ namespace detail {
 
 static void* LoadDLL(const Char* filePath) {
 #ifdef _WIN32
-    detail::GenericStaticallyLocalPath<TCHAR> localPath(filePath, StringHelper::strlen(filePath));
+    detail::GenericStaticallyLocalPath<TCHAR> localPath(filePath, static_cast<int>(StringHelper::strlen(filePath)));
     return ::LoadLibrary(localPath.c_str());
 #else
-    detail::GenericStaticallyLocalPath<char> localPath(filePath, StringHelper::strlen(filePath));
+    detail::GenericStaticallyLocalPath<char> localPath(filePath, static_cast<int>(StringHelper::strlen(filePath)));
     return dlopen(localPath.c_str(), RTLD_LAZY);
 #endif
 }
